[HLSL] Fold the D3D12 linear algebra shim back into HlslExecTestUtils.h
A separate LinAlgAbi.h has to be tracked by the packaging manifests
(vpacks and friends), which is a real cost for a block of code whose
whole purpose is to be deleted once the API ships in a released SDK.
Folding it into HlslExecTestUtils.h keeps the shim in one place, which
was the point of the earlier split, without adding a file.

The block sits immediately after the includes so the SDK glue stays
together, and is unchanged apart from comment wording that described it
as a file.

No functional change. WARP is unchanged at 24 total, 19 passed, 4
failed, 1 skipped, with the same non-passing set. The ASSERT_RUNTIME_*
blocks were re-verified in place: corrupting the FLOAT16 enumerator and
the LINEAR_ALGEBRA_SUPPORT feature id each fail the preview-SDK build at
the new location, so they still compare against the real d3d12.h rather
than against themselves.


Copilot-Session: 83725f5d-8e98-4c1d-91ee-ad47629e007b
