--- tools/clang/unittests/HLSLExec/LinAlgAbi.h
@@ -1,329 +1,0 @@
-///////////////////////////////////////////////////////////////////////////////
-//                                                                           //
-// LinAlgAbi.h                                                               //
-// Copyright (C) Microsoft Corporation. All rights reserved.                 //
-// This file is distributed under the University of Illinois Open Source     //
-// License. See LICENSE.TXT for details.                                     //
-//                                                                           //
-// Local copies of the D3D12 linear algebra API surface                      //
-//                                                                           //
-///////////////////////////////////////////////////////////////////////////////
-
-#ifndef LINALGABI_H
-#define LINALGABI_H
-
-#include <d3d12.h>
-
-// The released Windows SDK does not declare the linear algebra API yet, and
-// CheckFeatureSupport takes an untyped (void *, size) pair, so what this code
-// depends on is a runtime layout rather than a compile-time type. Everything
-// the tests need is therefore mirrored here.
-//
-// The whole shim lives in this one file so it can be deleted in one go. It is
-// transcribed verbatim from d3d12.h - same type names, same field names, same
-// enumerator names, same declaration order - so it can be diffed against the
-// header directly, and so removing it leaves the use sites needing only the
-// linalg_abi:: qualifier dropped.
-//
-// Enum-typed fields of the capability structures are declared as UINT because
-// the values crossing CheckFeatureSupport are opaque here. The enumerations are
-// the typed surface the tests use.
-//
-// None of it is taken on trust. When the preview SDK is present, the
-// ASSERT_RUNTIME_* blocks at the bottom of this file pin every value, size,
-// alignment and field offset to the real declarations. Those comparisons
-// qualify both sides explicitly because the names are deliberately identical:
-// unqualified lookup from inside the namespace would resolve to the local copy
-// twice and assert nothing.
-//
-// For the same reason there is no using-directive for this namespace, and there
-// must not be. The matrix-conversion helpers in HlslExecTestUtils.h take the
-// real D3D12 types, so pulling these names in unqualified would make every one
-// of those declarations ambiguous in a preview-SDK build.
-//
-// TODO: delete this file once the linear algebra API ships in a released
-// Windows SDK.
-namespace linalg_abi {
-
-// D3D12_FEATURE itself ships in the released SDK; only these two enumerators of
-// it are missing, so they are mirrored as constants rather than as an
-// enumeration.
-constexpr D3D12_FEATURE D3D12_FEATURE_LINEAR_ALGEBRA_SUPPORT =
-    static_cast<D3D12_FEATURE>(77);
-constexpr D3D12_FEATURE D3D12_FEATURE_LINEAR_ALGEBRA_MATRIX_OPERATION_SUPPORT =
-    static_cast<D3D12_FEATURE>(78);
-
-enum D3D12_LINEAR_ALGEBRA_TIER {
-  D3D12_LINEAR_ALGEBRA_TIER_NOT_SUPPORTED = 0,
-  D3D12_LINEAR_ALGEBRA_TIER_1_0 = 0x10
-};
-
-enum D3D12_LINEAR_ALGEBRA_DATATYPE {
-  D3D12_LINEAR_ALGEBRA_DATATYPE_NONE = 0,
-  D3D12_LINEAR_ALGEBRA_DATATYPE_SINT16 = 2,
-  D3D12_LINEAR_ALGEBRA_DATATYPE_UINT16 = 3,
-  D3D12_LINEAR_ALGEBRA_DATATYPE_SINT32 = 4,
-  D3D12_LINEAR_ALGEBRA_DATATYPE_UINT32 = 5,
-  D3D12_LINEAR_ALGEBRA_DATATYPE_FLOAT16 = 7,
-  D3D12_LINEAR_ALGEBRA_DATATYPE_FLOAT32 = 8,
-  D3D12_LINEAR_ALGEBRA_DATATYPE_SINT8 = 18,
-  D3D12_LINEAR_ALGEBRA_DATATYPE_UINT8 = 19,
-  D3D12_LINEAR_ALGEBRA_DATATYPE_FLOAT8_E4M3FN = 20,
-  D3D12_LINEAR_ALGEBRA_DATATYPE_FLOAT8_E5M2 = 21
-};
-
-enum D3D12_LINEAR_ALGEBRA_OPERATION_TYPE {
-  D3D12_LINEAR_ALGEBRA_OPERATION_TYPE_MATRIX_CONSTRUCTION = 0,
-  D3D12_LINEAR_ALGEBRA_OPERATION_TYPE_WAVE_MATRIX_MULTIPLY = 1,
-  D3D12_LINEAR_ALGEBRA_OPERATION_TYPE_THREADGROUP_MATRIX_MULTIPLY = 2,
-  D3D12_LINEAR_ALGEBRA_OPERATION_TYPE_THREAD_VECTOR_MATRIX_MULTIPLY = 3,
-  D3D12_LINEAR_ALGEBRA_OPERATION_TYPE_THREAD_OUTER_PRODUCT = 4,
-  D3D12_LINEAR_ALGEBRA_OPERATION_TYPE_ATOMIC_ACCUMULATE_STORE = 5
-};
-
-enum D3D12_LINEAR_ALGEBRA_MULTIPLICATION_SUPPORT_FLAGS {
-  D3D12_LINEAR_ALGEBRA_MULTIPLICATION_SUPPORT_FLAG_NONE = 0,
-  D3D12_LINEAR_ALGEBRA_MULTIPLICATION_SUPPORT_FLAG_SUPPORTED = 1,
-  D3D12_LINEAR_ALGEBRA_MULTIPLICATION_SUPPORT_FLAG_EMULATED_INPUTS = 2,
-  D3D12_LINEAR_ALGEBRA_MULTIPLICATION_SUPPORT_FLAG_EMULATED_OUTPUTS = 4,
-  D3D12_LINEAR_ALGEBRA_MULTIPLICATION_SUPPORT_FLAG_TRANSPOSE = 8
-};
-
-struct D3D12_LINEAR_ALGEBRA_MATRIX_SHAPE {
-  UINT M;
-  UINT K;
-  UINT N;
-};
-
-typedef D3D12_LINEAR_ALGEBRA_MATRIX_SHAPE
-    D3D12_LINEAR_ALGEBRA_MATRIX_MULTIPLY_SHAPE;
-
-struct D3D12_FEATURE_DATA_LINEAR_ALGEBRA_SUPPORT {
-  UINT LinearAlgebraTier;
-};
-
-struct D3D12_LINEAR_ALGEBRA_MATRIX_CONSTRUCTION_SUPPORT {
-  UINT ComponentType;
-  UINT WaveSize;
-  D3D12_LINEAR_ALGEBRA_MATRIX_SHAPE Shape;
-  BOOL Supported;
-};
-
-struct D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_INPUTS {
-  UINT WaveSize;
-  UINT MatrixAComponentType;
-  UINT MatrixBComponentType;
-  UINT AccumulatorComponentType;
-};
-
-struct D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_SUPPORT {
-  D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_INPUTS Inputs;
-  D3D12_LINEAR_ALGEBRA_MATRIX_SHAPE Shape;
-  UINT SupportFlags;
-};
-
-struct D3D12_LINEAR_ALGEBRA_THREADGROUP_MATRIX_MULTIPLY_SUPPORT {
-  D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_INPUTS WaveInputs;
-  D3D12_LINEAR_ALGEBRA_MATRIX_SHAPE Shape;
-  UINT SupportFlags;
-  UINT MinThreadGroupSize;
-  UINT MaxThreadGroupSize;
-  UINT PreferredThreadGroupSize;
-};
-
-struct D3D12_LINEAR_ALGEBRA_THREAD_VECTOR_MATRIX_MULTIPLY_SUPPORT {
-  UINT VectorInputType;
-  UINT MatrixInputType;
-  UINT BiasInputType;
-  UINT VectorResultType;
-  UINT SupportFlags;
-};
-
-struct D3D12_LINEAR_ALGEBRA_THREAD_OUTER_PRODUCT_SUPPORT {
-  UINT InputComponentType;
-  UINT ResultComponentType;
-  BOOL Supported;
-};
-
-struct D3D12_LINEAR_ALGEBRA_ATOMIC_ACCUMULATE_STORE_SUPPORT {
-  UINT ComponentType;
-  BOOL RWByteAddressBufferSupported;
-  BOOL GroupSharedSupported;
-};
-
-struct D3D12_FEATURE_DATA_LINEAR_ALGEBRA_MATRIX_OPERATION_SUPPORT {
-  UINT OperationType;
-  union {
-    D3D12_LINEAR_ALGEBRA_MATRIX_CONSTRUCTION_SUPPORT MatrixConstruction;
-    D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_SUPPORT WaveMatrixMultiply;
-    D3D12_LINEAR_ALGEBRA_THREADGROUP_MATRIX_MULTIPLY_SUPPORT
-    ThreadGroupMatrixMultiply;
-    D3D12_LINEAR_ALGEBRA_THREAD_VECTOR_MATRIX_MULTIPLY_SUPPORT
-    ThreadVectorMatrixMultiply;
-    D3D12_LINEAR_ALGEBRA_THREAD_OUTER_PRODUCT_SUPPORT ThreadOuterProductSupport;
-    D3D12_LINEAR_ALGEBRA_ATOMIC_ACCUMULATE_STORE_SUPPORT AccumulateStore;
-  };
-};
-
-} // namespace linalg_abi
-
-// Everything above is pinned to the real declarations here. This only compiles
-// against a preview SDK, so it is not reached by an ordinary build; run a
-// preview-SDK compile whenever the copies above are touched.
-#if defined(DIRECT3D_LINEAR_ALGEBRA)
-
-#define ASSERT_RUNTIME_ENUM(EnumeratorName)                                    \
-  static_assert(static_cast<UINT>(linalg_abi::EnumeratorName) ==               \
-                    static_cast<UINT>(::EnumeratorName),                       \
-                "Linear algebra runtime enum value changed")
-
-ASSERT_RUNTIME_ENUM(D3D12_FEATURE_LINEAR_ALGEBRA_SUPPORT);
-ASSERT_RUNTIME_ENUM(D3D12_FEATURE_LINEAR_ALGEBRA_MATRIX_OPERATION_SUPPORT);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_TIER_NOT_SUPPORTED);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_TIER_1_0);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_DATATYPE_NONE);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_DATATYPE_SINT16);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_DATATYPE_UINT16);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_DATATYPE_SINT32);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_DATATYPE_UINT32);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_DATATYPE_FLOAT16);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_DATATYPE_FLOAT32);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_DATATYPE_SINT8);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_DATATYPE_UINT8);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_DATATYPE_FLOAT8_E4M3FN);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_DATATYPE_FLOAT8_E5M2);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_OPERATION_TYPE_MATRIX_CONSTRUCTION);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_OPERATION_TYPE_WAVE_MATRIX_MULTIPLY);
-ASSERT_RUNTIME_ENUM(
-    D3D12_LINEAR_ALGEBRA_OPERATION_TYPE_THREADGROUP_MATRIX_MULTIPLY);
-ASSERT_RUNTIME_ENUM(
-    D3D12_LINEAR_ALGEBRA_OPERATION_TYPE_THREAD_VECTOR_MATRIX_MULTIPLY);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_OPERATION_TYPE_THREAD_OUTER_PRODUCT);
-ASSERT_RUNTIME_ENUM(
-    D3D12_LINEAR_ALGEBRA_OPERATION_TYPE_ATOMIC_ACCUMULATE_STORE);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_MULTIPLICATION_SUPPORT_FLAG_NONE);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_MULTIPLICATION_SUPPORT_FLAG_SUPPORTED);
-ASSERT_RUNTIME_ENUM(
-    D3D12_LINEAR_ALGEBRA_MULTIPLICATION_SUPPORT_FLAG_EMULATED_INPUTS);
-ASSERT_RUNTIME_ENUM(
-    D3D12_LINEAR_ALGEBRA_MULTIPLICATION_SUPPORT_FLAG_EMULATED_OUTPUTS);
-ASSERT_RUNTIME_ENUM(D3D12_LINEAR_ALGEBRA_MULTIPLICATION_SUPPORT_FLAG_TRANSPOSE);
-
-#undef ASSERT_RUNTIME_ENUM
-
-// The local copies share their names with the real header types, so these
-// comparisons must be explicitly qualified: linalg_abi:: for the local copy and
-// :: for the SDK declaration. Unqualified lookup from inside this namespace
-// resolves to the local copy on both sides and would assert nothing.
-#define ASSERT_RUNTIME_ABI(TypeName)                                           \
-  static_assert(sizeof(linalg_abi::TypeName) == sizeof(::TypeName),            \
-                "Linear algebra runtime ABI size changed");                    \
-  static_assert(alignof(linalg_abi::TypeName) == alignof(::TypeName),          \
-                "Linear algebra runtime ABI alignment changed")
-
-ASSERT_RUNTIME_ABI(D3D12_FEATURE_DATA_LINEAR_ALGEBRA_SUPPORT);
-ASSERT_RUNTIME_ABI(D3D12_LINEAR_ALGEBRA_MATRIX_CONSTRUCTION_SUPPORT);
-ASSERT_RUNTIME_ABI(D3D12_LINEAR_ALGEBRA_MATRIX_MULTIPLY_SHAPE);
-ASSERT_RUNTIME_ABI(D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_INPUTS);
-ASSERT_RUNTIME_ABI(D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_SUPPORT);
-ASSERT_RUNTIME_ABI(D3D12_LINEAR_ALGEBRA_THREADGROUP_MATRIX_MULTIPLY_SUPPORT);
-ASSERT_RUNTIME_ABI(D3D12_LINEAR_ALGEBRA_THREAD_VECTOR_MATRIX_MULTIPLY_SUPPORT);
-ASSERT_RUNTIME_ABI(D3D12_LINEAR_ALGEBRA_THREAD_OUTER_PRODUCT_SUPPORT);
-ASSERT_RUNTIME_ABI(D3D12_LINEAR_ALGEBRA_ATOMIC_ACCUMULATE_STORE_SUPPORT);
-ASSERT_RUNTIME_ABI(D3D12_FEATURE_DATA_LINEAR_ALGEBRA_MATRIX_OPERATION_SUPPORT);
-
-#undef ASSERT_RUNTIME_ABI
-
-#define ASSERT_RUNTIME_OFFSET(TypeName, Field)                                 \
-  static_assert(offsetof(linalg_abi::TypeName, Field) ==                       \
-                    offsetof(::TypeName, Field),                               \
-                "Linear algebra runtime ABI field offset changed")
-
-ASSERT_RUNTIME_OFFSET(D3D12_FEATURE_DATA_LINEAR_ALGEBRA_SUPPORT,
-                      LinearAlgebraTier);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_MATRIX_CONSTRUCTION_SUPPORT,
-                      ComponentType);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_MATRIX_CONSTRUCTION_SUPPORT,
-                      WaveSize);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_MATRIX_CONSTRUCTION_SUPPORT, Shape);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_MATRIX_CONSTRUCTION_SUPPORT,
-                      Supported);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_MATRIX_MULTIPLY_SHAPE, M);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_MATRIX_MULTIPLY_SHAPE, K);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_MATRIX_MULTIPLY_SHAPE, N);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_INPUTS,
-                      WaveSize);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_INPUTS,
-                      MatrixAComponentType);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_INPUTS,
-                      MatrixBComponentType);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_INPUTS,
-                      AccumulatorComponentType);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_SUPPORT,
-                      Inputs);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_SUPPORT, Shape);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_WAVE_MATRIX_MULTIPLY_SUPPORT,
-                      SupportFlags);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_THREADGROUP_MATRIX_MULTIPLY_SUPPORT,
-                      WaveInputs);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_THREADGROUP_MATRIX_MULTIPLY_SUPPORT,
-                      Shape);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_THREADGROUP_MATRIX_MULTIPLY_SUPPORT,
-                      SupportFlags);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_THREADGROUP_MATRIX_MULTIPLY_SUPPORT,
-                      MinThreadGroupSize);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_THREADGROUP_MATRIX_MULTIPLY_SUPPORT,
-                      MaxThreadGroupSize);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_THREADGROUP_MATRIX_MULTIPLY_SUPPORT,
-                      PreferredThreadGroupSize);
-ASSERT_RUNTIME_OFFSET(
-    D3D12_LINEAR_ALGEBRA_THREAD_VECTOR_MATRIX_MULTIPLY_SUPPORT,
-    VectorInputType);
-ASSERT_RUNTIME_OFFSET(
-    D3D12_LINEAR_ALGEBRA_THREAD_VECTOR_MATRIX_MULTIPLY_SUPPORT,
-    MatrixInputType);
-ASSERT_RUNTIME_OFFSET(
-    D3D12_LINEAR_ALGEBRA_THREAD_VECTOR_MATRIX_MULTIPLY_SUPPORT, BiasInputType);
-ASSERT_RUNTIME_OFFSET(
-    D3D12_LINEAR_ALGEBRA_THREAD_VECTOR_MATRIX_MULTIPLY_SUPPORT,
-    VectorResultType);
-ASSERT_RUNTIME_OFFSET(
-    D3D12_LINEAR_ALGEBRA_THREAD_VECTOR_MATRIX_MULTIPLY_SUPPORT, SupportFlags);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_THREAD_OUTER_PRODUCT_SUPPORT,
-                      InputComponentType);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_THREAD_OUTER_PRODUCT_SUPPORT,
-                      ResultComponentType);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_THREAD_OUTER_PRODUCT_SUPPORT,
-                      Supported);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_ATOMIC_ACCUMULATE_STORE_SUPPORT,
-                      ComponentType);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_ATOMIC_ACCUMULATE_STORE_SUPPORT,
-                      RWByteAddressBufferSupported);
-ASSERT_RUNTIME_OFFSET(D3D12_LINEAR_ALGEBRA_ATOMIC_ACCUMULATE_STORE_SUPPORT,
-                      GroupSharedSupported);
-ASSERT_RUNTIME_OFFSET(
-    D3D12_FEATURE_DATA_LINEAR_ALGEBRA_MATRIX_OPERATION_SUPPORT, OperationType);
-ASSERT_RUNTIME_OFFSET(
-    D3D12_FEATURE_DATA_LINEAR_ALGEBRA_MATRIX_OPERATION_SUPPORT,
-    MatrixConstruction);
-ASSERT_RUNTIME_OFFSET(
-    D3D12_FEATURE_DATA_LINEAR_ALGEBRA_MATRIX_OPERATION_SUPPORT,
-    WaveMatrixMultiply);
-ASSERT_RUNTIME_OFFSET(
-    D3D12_FEATURE_DATA_LINEAR_ALGEBRA_MATRIX_OPERATION_SUPPORT,
-    ThreadGroupMatrixMultiply);
-ASSERT_RUNTIME_OFFSET(
-    D3D12_FEATURE_DATA_LINEAR_ALGEBRA_MATRIX_OPERATION_SUPPORT,
-    ThreadVectorMatrixMultiply);
-ASSERT_RUNTIME_OFFSET(
-    D3D12_FEATURE_DATA_LINEAR_ALGEBRA_MATRIX_OPERATION_SUPPORT,
-    ThreadOuterProductSupport);
-ASSERT_RUNTIME_OFFSET(
-    D3D12_FEATURE_DATA_LINEAR_ALGEBRA_MATRIX_OPERATION_SUPPORT,
-    AccumulateStore);
-
-#undef ASSERT_RUNTIME_OFFSET
-
-#endif // defined(DIRECT3D_LINEAR_ALGEBRA)
-
-#endif // LINALGABI_H
